Stabilize Road and Train satellite maps
Keep zoom tied to explicit slippy-map levels instead of jumping through arbitrary camera limits, preserve the previous map while replacement tiles load, retry rendering after tile failures, remove composite seams, and center authoring maps on project terrain bounds or a land-based default.

+++ src/engine/lanemaker/engine/map_view_gl.h
@@ -224,7 +224,7 @@ namespace LM
         void compositeMapImage();
         static void latLonToTile(double lat, double lon, int z, int& tx, int& ty);
         static void tileToLatLon(int tx, int ty, int z, double& lat, double& lon);
-        static double metersPerPixel(double lat, int z);
+        static double metersPerPixel(double lat, double z);
 	};
 
 	extern MapViewGL* g_mapViewGL;
